Generalize remove/find to work on vector-like types with unsigned size.

+++ minisat/mtl/Alg.h
@@ -36,9 +36,9 @@ template<class V, class T>
 static inline void remove(V& ts, const T& t)
 {
     int j = 0;
-    for (; j < ts.size() && ts[j] != t; j++);
-    assert(j < ts.size());
-    for (; j < ts.size()-1; j++) ts[j] = ts[j+1];
+    for (; j < (int)ts.size() && ts[j] != t; j++);
+    assert(j < (int)ts.size());
+    for (; j < (int)ts.size()-1; j++) ts[j] = ts[j+1];
     ts.pop();
 }
 
@@ -47,8 +47,8 @@ template<class V, class T>
 static inline bool find(V& ts, const T& t)
 {
     int j = 0;
-    for (; j < ts.size() && ts[j] != t; j++);
-    return j < ts.size();
+    for (; j < (int)ts.size() && ts[j] != t; j++);
+    return j < (int)ts.size();
 }
